fix #7 : of .emit on augmentedfunctions
I added performance.js to benchmark which of .apply, .call and .bind
would be most proper to use in this situation. Given that on and emit
are both simple enough to be implemented with either of the three, we
might as well pick the faster one. This code is going to be a very hot,
so even the slightest optimizations will probably be effective.